pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "houdini-mcp"
version = "0.1.0"
description = "MCP Server for SideFX Houdini — control Houdini via RPyC through the Model Context Protocol"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
keywords = ["houdini", "mcp", "rpyc", "3d", "vfx", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Multimedia :: Graphics :: 3D Rendering",
    "Programming Language :: Python :: 3.11",
    "Operating System :: Microsoft :: Windows",
]
dependencies = [
    "fastmcp>=2.0",
    "rpyc>=4.1,<5",
    "Pillow>=10.0",
    "scikit-image>=0.22",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
houdini-mcp = "houdini_mcp.__main__:main"
